backend/requirements.txt
django
django-rest-framework
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt